# trie: reduce allocations in stacktrie