--- a/trie/encoding.go
+++ b/trie/encoding.go
@@ -107,12 +107,13 @@ func keybytesToHex(str []byte) []byte {
 // writeHexKey writes the hexkey into the given slice.
 // OBS! This method omits the termination flag.
 // OBS! The dst slice must be at least 2x as large as the key
-func writeHexKey(dst []byte, key []byte) {
+func writeHexKey(dst []byte, key []byte) []byte {
 	_ = dst[2*len(key)-1]
 	for i, b := range key {
 		dst[i*2] = b / 16
 		dst[i*2+1] = b % 16
 	}
+	return dst[:2*len(key)]
 }
 
 // hexToKeybytes turns hex nibbles into key bytes.