--- a/trie/node_enc.go
+++ b/trie/node_enc.go
@@ -81,12 +81,6 @@ func (n *extNodeEncoder) encode(w rlp.EncoderBuffer) {
 
 func (n *leafNodeEncoder) encode(w rlp.EncoderBuffer) {
 	offset := w.List()
-
-	// Encode the key to Compact format. The assumption is held
-	// that it's safe to modify the key in place.
-	n.Key = append(n.Key, 16) // termination flag
-	n.Key = hexToCompactInPlace(n.Key)
-
 	w.WriteBytes(n.Key) // Compact format key
 	w.WriteBytes(n.Val) // Value node, must be non-nil
 	w.ListEnd(offset)