… trie/stacktrie_test.go