@holiman 2024-11-11
Update
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/trie
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
             │ stacktrie.2  │              stacktrie.3              │
             │    sec/op    │    sec/op     vs base                 │
Insert100K-8   78.51m ± ∞ ¹   69.50m ± 12%  -11.47% (p=0.010 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

             │  stacktrie.2  │              stacktrie.3              │
             │     B/op      │     B/op      vs base                 │
Insert100K-8   6.931Mi ± ∞ ¹   4.640Mi ± 0%  -33.06% (p=0.003 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

             │ stacktrie.2  │         stacktrie.3          │
             │  allocs/op   │  allocs/op   vs base         │
Insert100K-8   326.7k ± ∞ ¹   226.7k ± 0%  -30.61% (n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95
```
Total
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/trie
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
             │ stacktrie.1  │              stacktrie.3              │
             │    sec/op    │    sec/op     vs base                 │
Insert100K-8   94.80m ± ∞ ¹   69.50m ± 12%  -26.69% (p=0.003 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

             │  stacktrie.1   │              stacktrie.3              │
             │      B/op      │     B/op      vs base                 │
Insert100K-8   12.890Mi ± ∞ ¹   4.640Mi ± 0%  -64.01% (p=0.003 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

             │ stacktrie.1  │         stacktrie.3          │
             │  allocs/op   │  allocs/op   vs base         │
Insert100K-8   553.4k ± ∞ ¹   226.7k ± 0%  -59.03% (n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

```