@holiman 2024-11-11
New progress

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/trie
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
             │ stacktrie.3  │          stacktrie.4          │
             │    sec/op    │    sec/op     vs base         │
Insert100K-8   69.50m ± 12%   74.59m ± 14%  ~ (p=0.128 n=7)

             │ stacktrie.3  │             stacktrie.4             │
             │     B/op     │     B/op      vs base               │
Insert100K-8   4.640Mi ± 0%   3.112Mi ± 0%  -32.93% (p=0.001 n=7)

             │ stacktrie.3 │            stacktrie.4             │
             │  allocs/op  │  allocs/op   vs base               │
Insert100K-8   226.7k ± 0%   126.7k ± 0%  -44.11% (p=0.001 n=7)
```

total
```
$ benchstat stacktrie.1 stacktrie.4
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/trie
cpu: 12th Gen Intel(R) Core(TM) i7-1270P
             │ stacktrie.1  │              stacktrie.4              │
             │    sec/op    │    sec/op     vs base                 │
Insert100K-8   94.80m ± ∞ ¹   74.59m ± 14%  -21.32% (p=0.003 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

             │  stacktrie.1   │              stacktrie.4              │
             │      B/op      │     B/op      vs base                 │
Insert100K-8   12.890Mi ± ∞ ¹   3.112Mi ± 0%  -75.86% (p=0.003 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95

             │ stacktrie.1  │             stacktrie.4              │
             │  allocs/op   │  allocs/op   vs base                 │
Insert100K-8   553.4k ± ∞ ¹   126.7k ± 0%  -77.10% (p=0.003 n=5+7)
¹ need >= 6 samples for confidence interval at level 0.95
```