@holiman 2024-11-11
Argh only this little thing left: returning a slice to a pool somehow leaks something. I guess somehow I expose an underlying array, or at least make the compiler think so, hence it reallocs the slice when I return it to the pool.  

![Screenshot 2024-11-11 at 22-17-15 trie test alloc_space](https://github.com/user-attachments/assets/949e5d0a-0536-4d76-97dd-b8206e0fd872)
![Screenshot 2024-11-11 at 22-18-40 trie test alloc_space](https://github.com/user-attachments/assets/522ab0fa-08a3-46b0-a41a-cd132edd4c1c)