@rjl493456442 2024-11-21
@holiman I think we should compare the relevant metrics during the snap sync and to see the overall impact brought by this change.