@holiman 2024-11-25
Running a snapsync benchmarkers now, using `partialwipe` (ancients intact), 

- master on `bench06`
- This PR on `bench05`