@holiman 2024-11-26
So, here are some charts. The first third of the charts is the snapsync, the latter two thirds are snap-gen + block by block sync. The yellow (this PR) finished slightly faster (13 minutes or so). 

![Screenshot 2024-11-26 at 12-18-25 Dual Geth - Grafana](https://github.com/user-attachments/assets/66f7b317-0913-4a7a-9771-9a2b3b35de88)

This is despite that the `05` (this PR) struggled against larger iowait. Perhaps it was bottlenecked by the disk? 

![Screenshot 2024-11-26 at 12-19-55 Dual Geth - Grafana](https://github.com/user-attachments/assets/d94ffeb2-1609-48a6-b0bf-59416c891f03)

The `05` had some block downloading to do in the beginnig, but after that the ingress rates were fairly equal: 

![Screenshot 2024-11-26 at 12-21-30 Dual Geth - Grafana](https://github.com/user-attachments/assets/ab5ed6ed-c686-47ba-a8ba-56861b01cf44)

Memory charts doesn't show much, except that _maybe_ the lead that `05` had after sync was increased further during generation, to about `25 minutes`

![Screenshot 2024-11-26 at 12-22-45 Dual Geth - Grafana](https://github.com/user-attachments/assets/e0f9e4d3-90e2-4e07-a09b-e3f33ef3b2a5)