@holiman 2025-01-14
Also worth noting: https://github.com/ethereum/go-ethereum/pull/30747  . Whereas _this_ PR takes ownership of `st.val` IFF the node is converted into a hashedNode, the other PR takes full ownership of `st.val` immediately as it enters the stacktrie. 

Which means that it's less elaborate: the stacktrie always owns the `st.val`, we don't need thorough analysis to realise the safety of it. OTOH, that approach will always copy-on-ingress. For `DeriveSha`, the that approach is objectively better, since the copy-on-ingress will copy to a pool-managed slice, the outer caller can safely reuse it's own slice.