@holiman 2025-01-14
> 100K is a big number for a benchmark, but the scale of what the stacktrie is being used for is 10^4 times that.

Actually, for the most part, we use the stacktrie to validate sequences of accounts / slots during snap sync. These are typically on the order of 10K elements, IIRC. 

We don't typically feed the entire trie into a stacktrie: you'd only do that if you want to do some of the deep verifications. 

When we generate snapshot data from the trie data, I'm not sure the stacktrie is involved, really.